requirements.txt
python-telegram-bot==20.7
selenium==4.21.0
python-dotenv
ta
pandas